Move the backend proxy to a concern
and document it a bit...
